+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- `Stats.progress`: Estimate how much of the log data has been parsed so far.
+
 ## [0.3.0] - 2023-03-01
 
 ### Added